backend/requirements.txt
# ESP32 Wildlife Camera Backend Requirements
# Core Flask framework and extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Task queue
celery[redis]==5.3.2
redis==5.0.0

# Image processing
Pillow==10.0.1
opencv-python-headless==4.8.1.78

# Machine Learning
tensorflow==2.13.0
numpy==1.24.3
scikit-learn==1.3.0

# HTTP and API
requests==2.31.0
urllib3==2.0.4

# Data processing
pandas==2.1.1

# Cloud storage (optional)
boto3==1.28.62
google-cloud-storage==2.10.0

# Monitoring and logging
gunicorn==21.2.0
python-dotenv==1.0.0

# Development and testing
pytest==7.4.2
pytest-flask==1.2.0
black==23.9.1
flake8==6.1.0

# Security
cryptography==41.0.4
bcrypt==4.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3